+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Blocks : display blocks are now more configurable (custom selector and type)
 - Schemas refactor: add support of Ellipsed tuple (Tuple[T, ...])